Add Material Design styling
This commit adds some styling that implements the Material Design philosophy by utilizing the metaphor of elevation of a card of paper under ambient lighting. When a user hovers over a card containing the information of a single customer, the card appears to elevate, which visually and aesthetically indicates the user's interaction with it.