Fix a routing bug in production
This commit fixes a bug in the production environment that prevented a URL for a single customer from working.

The prior implementation relied on there being no numbers in the URL besides the customer's primary key, but this didn't work in production because the AWS S3 URL has a few other numbers in it. With the current commit applied, the URL for a single customer will work as long as the customer's primary key is the last number in the URL. So, an example of a functioning URL is: http://consuming-ecommerce-rest-api.s3-website.us-east-2.amazonaws.com/#customers/1