Fix production routing bug for all customers
The prior commit implemented a fix for a production bug in visiting a URL for a single customer, but the fix only worked when the customer's primary key was a single digit. With the current commit applied, URLs for a single customer work for every customer, no matter the size of the customer's primary key.